move checks query into pr shared
